add fr_strlen size_t

+++ libft.h
@@ -6,18 +6,20 @@
 /*   By: asobrino <asobrino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 09:59:48 by asobrino          #+#    #+#             */
-/*   Updated: 2025/04/10 13:18:45 by asobrino         ###   ########.fr       */
+/*   Updated: 2025/04/10 13:30:05 by asobrino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-int					ft_isalpha(int c);
-int					ft_isdigit(int c);
-int					ft_isalnum(int c);
-int					ft_isascii(unsigned char c);
-int					ft_isprint(int c);
-long unsigned int	ft_strlen(const char *s);
+# include <unistd.h>
+
+int		ft_isalpha(int c);
+int		ft_isdigit(int c);
+int		ft_isalnum(int c);
+int		ft_isascii(unsigned char c);
+int		ft_isprint(int c);
+size_t	ft_strlen(const char *s);
 
 #endif
